+++ src/opera_disp_tms/create_tile_map.py
@@ -11,18 +11,28 @@ from osgeo import gdal, gdalconst, osr
 gdal.UseExceptions()
 
 
-def get_tile_extent(info: dict, output_folder: Path) -> None:
-    """Generate file with the bounds of the newly created vrt
-        Will return a file with: {"extent": [minx, miny, maxx, maxy], "EPSG": %EPSG}
+def create_bounds_file(info: dict, scale_range: list, output_folder: Path) -> None:
+    """Generate file with the bounds and scale ranges of the newly created vrt
 
     Args:
         info: gdalinfo dict from vrt file
+        scale_range: list with min and max of tile map
         output_folder: folder to write "extent.json"
     """
     minx, miny = info['cornerCoordinates']['lowerLeft']
     maxx, maxy = info['cornerCoordinates']['upperRight']
     proj = osr.SpatialReference(info['coordinateSystem']['wkt'])
-    extent = {'extent': [minx, miny, maxx, maxy], 'EPSG': int(proj.GetAttrValue('AUTHORITY', 1))}
+
+    sig_figs = 3
+
+    extent = {
+        'extent': [minx, miny, maxx, maxy],
+        'EPSG': int(proj.GetAttrValue('AUTHORITY', 1)),
+        'scale_range': {
+            'range': [round(scale, sig_figs) for scale in scale_range],
+            'units': 'm/yr',
+        },
+    }
 
     if not output_folder.exists():
         output_folder.mkdir()
@@ -49,7 +59,7 @@ def create_tile_map(output_folder: str, input_rasters: list[str], scale_range: l
         stats = vrt_info['bands'][0]['metadata']['']
 
         if scale_range is None:
-            scale_range = [stats['STATISTICS_MINIMUM'], stats['STATISTICS_MAXIMUM']]
+            scale_range = [float(stats['STATISTICS_MINIMUM']), float(stats['STATISTICS_MAXIMUM'])]
 
         gdal.Translate(
             destName=byte_vrt.name,
@@ -74,7 +84,7 @@ def create_tile_map(output_folder: str, input_rasters: list[str], scale_range: l
         subprocess.run(command)
 
         # get bounds of VRT and write to file
-        get_tile_extent(vrt_info, Path(output_folder))
+        create_bounds_file(vrt_info, scale_range, Path(output_folder))
 
 
 def main():