+++ src/opera_disp_tms/utils.py
@@ -66,6 +66,10 @@ def upload_file_to_s3(path_to_file: Path, bucket: str, key):
     extra_args = {'ContentType': guess_type(path_to_file)[0]}
     S3_CLIENT.upload_file(str(path_to_file), bucket, key, ExtraArgs=extra_args)
 
+    # tag files as 'product' so hyp3 doesn't treat the .png files as browse images
+    tag_set = {'TagSet': [{'Key': 'file_type', 'Value': 'product'}]}
+    S3_CLIENT.put_object_tagging(Bucket=bucket, Key=key, Tagging=tag_set)
+
 
 def upload_dir_to_s3(path_to_dir: Path, bucket: str, prefix: str = ''):
     """Upload a local directory, subdirectory, and all contents to an S3 bucket