+++ tests/test_prep_stack.py
@@ -19,17 +19,15 @@ def test_find_needed_granules():
     ]
     fn_name = 'opera_disp_tms.prep_stack.find_granules_for_frame'
     with mock.patch(fn_name, return_value=granules):
-        needed_granules = prep_stack.find_needed_granules(
-            [1], datetime(2021, 1, 1), datetime(2021, 1, 9), strategy='all'
-        )
-        assert len(needed_granules[1]) == 3
-        assert needed_granules == granules[:3]
+        needed_granules = prep_stack.find_needed_granules(1, datetime(2021, 1, 1), datetime(2021, 1, 9), strategy='all')
+        assert len(needed_granules) == 3
+        assert needed_granules == granules[:3]  # type: ignore[comparison-overlap]
 
         needed_granules = prep_stack.find_needed_granules(
-            [1], datetime(2021, 1, 1), datetime(2021, 1, 9), strategy='spanning'
+            1, datetime(2021, 1, 1), datetime(2021, 1, 9), strategy='spanning'
         )
         assert len(needed_granules) == 2
-        assert needed_granules == granules[1:3]
+        assert needed_granules == granules[1:3]  # type: ignore[comparison-overlap]
 
 
 def test_restrict_to_spanning_set():
@@ -39,17 +37,17 @@ def test_restrict_to_spanning_set():
         GranuleStub(1, datetime(2021, 1, 4), datetime(2021, 1, 7)),
         GranuleStub(1, datetime(2021, 1, 7), datetime(2021, 1, 10)),
     ]
-    result = prep_stack.restrict_to_spanning_set(granules)
+    result = prep_stack.restrict_to_spanning_set(granules)  # type: ignore[arg-type]
     assert len(result) == 3
-    assert result == granules
+    assert result == granules  # type: ignore[comparison-overlap]
 
     granules.append(GranuleStub(1, datetime(2021, 1, 7), datetime(2021, 1, 13)))
-    result = prep_stack.restrict_to_spanning_set(granules)
+    result = prep_stack.restrict_to_spanning_set(granules)  # type: ignore[arg-type]
     assert len(result) == 3
-    assert result == [granules[0], granules[1], granules[3]]
+    assert result == [granules[0], granules[1], granules[3]]  # type: ignore[comparison-overlap]
 
     with pytest.raises(ValueError, match='Granules do not form a spanning set.'):
-        prep_stack.restrict_to_spanning_set([granules[0], granules[2], granules[3]])
+        prep_stack.restrict_to_spanning_set([granules[0], granules[2], granules[3]])  # type: ignore[list-item]
 
 
 def test_align_to_common_reference_date():