+++ tests/test_utils.py
@@ -40,6 +40,7 @@ def test_upload_file_to_s3(tmp_path, s3_stubber):
         'Tagging': {'TagSet': [{'Key': 'file_type', 'Value': 'product'}]},
     }
     s3_stubber.add_response(method='put_object', expected_params=expected_params, service_response={})
+    s3_stubber.add_response(method='put_object_tagging', expected_params=tag_params, service_response={})
 
     file_to_upload = tmp_path / 'myFile.png'
     file_to_upload.touch()